[#IOPID-857] WIP: little refactor after rebase
