+++ LockSession/function.json
@@ -1,7 +1,7 @@
 {
   "bindings": [
     {
-      "authLevel": "anonymous",
+      "authLevel": "function",
       "type": "httpTrigger",
       "direction": "in",
       "name": "req",
@@ -17,5 +17,5 @@
     }
   ],
   "scriptFile": "../dist/main.js",
-  "entryPoint": "Info"
+  "entryPoint": "LockSession"
 }
