+++ src/utils/ioweb/dependency.ts
@@ -1,3 +1,9 @@
+import { Client } from "../../generated/definitions/backend-internal/client";
+
 export type FnLockSessionClientDependency = {
-  backendInternalClient: any;
+  readonly backendInternalClient: Client<"token">;
+};
+
+export type LogoutDependencies = {
+  readonly backendInternalClient: Client<"token">;
 };
