--- src/utils/logout/dependency.ts
@@ -1,4 +1,0 @@
-import { Client } from "../../generated/definitions/backend-internal/client";
-export type LogoutDependencies = {
-  readonly backendInternalClient: Client<"token">;
-};
